+++ sdk/communication/azure-communication-callautomation/src/main/java/com/azure/communication/callautomation/implementation/models/UnmuteParticipantsRequestInternal.java
@@ -5,7 +5,6 @@
 package com.azure.communication.callautomation.implementation.models;
 
 import com.azure.core.annotation.Fluent;
-import com.azure.core.annotation.Generated;
 import com.azure.json.JsonReader;
 import com.azure.json.JsonSerializable;
 import com.azure.json.JsonToken;
@@ -22,19 +21,16 @@ public final class UnmuteParticipantsRequestInternal implements JsonSerializable
      * Participants to be unmuted from the call.
      * Only ACS Users are supported.
      */
-    @Generated
     private List<CommunicationIdentifierModel> targetParticipants;
 
     /*
      * Used by customers when calling mid-call actions to correlate the request to the response event.
      */
-    @Generated
     private String operationContext;
 
     /**
      * Creates an instance of UnmuteParticipantsRequestInternal class.
      */
-    @Generated
     public UnmuteParticipantsRequestInternal() {
     }
 
@@ -44,7 +40,6 @@ public final class UnmuteParticipantsRequestInternal implements JsonSerializable
      * 
      * @return the targetParticipants value.
      */
-    @Generated
     public List<CommunicationIdentifierModel> getTargetParticipants() {
         return this.targetParticipants;
     }
@@ -56,7 +51,6 @@ public final class UnmuteParticipantsRequestInternal implements JsonSerializable
      * @param targetParticipants the targetParticipants value to set.
      * @return the UnmuteParticipantsRequestInternal object itself.
      */
-    @Generated
     public UnmuteParticipantsRequestInternal
         setTargetParticipants(List<CommunicationIdentifierModel> targetParticipants) {
         this.targetParticipants = targetParticipants;
@@ -69,7 +63,6 @@ public final class UnmuteParticipantsRequestInternal implements JsonSerializable
      * 
      * @return the operationContext value.
      */
-    @Generated
     public String getOperationContext() {
         return this.operationContext;
     }
@@ -81,7 +74,6 @@ public final class UnmuteParticipantsRequestInternal implements JsonSerializable
      * @param operationContext the operationContext value to set.
      * @return the UnmuteParticipantsRequestInternal object itself.
      */
-    @Generated
     public UnmuteParticipantsRequestInternal setOperationContext(String operationContext) {
         this.operationContext = operationContext;
         return this;
@@ -90,7 +82,6 @@ public final class UnmuteParticipantsRequestInternal implements JsonSerializable
     /**
      * {@inheritDoc}
      */
-    @Generated
     @Override
     public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
         jsonWriter.writeStartObject();
@@ -109,7 +100,6 @@ public final class UnmuteParticipantsRequestInternal implements JsonSerializable
      * @throws IllegalStateException If the deserialized JSON object was missing any required properties.
      * @throws IOException If an error occurs while reading the UnmuteParticipantsRequestInternal.
      */
-    @Generated
     public static UnmuteParticipantsRequestInternal fromJson(JsonReader jsonReader) throws IOException {
         return jsonReader.readObject(reader -> {
             UnmuteParticipantsRequestInternal deserializedUnmuteParticipantsRequestInternal